+++ CHANGELOG.rst
@@ -22,6 +22,8 @@ New features and notable changes:
 - Pygments required >= 2.13.0. (:issue:`799`)
 - Add a second theme for HTML report inspired by GitHub. (:issue:`793`)
 - Add :option:`--fail-under-decision` and :option:`--fail-under-function` which will error under a given minimum coverage. (:issue:`773`)
+- Add function coverage to data model. (:issue:`822`)
+- Add support for importing Cobertura XML files with ``--cobertura-add-tracefile`` option. (:issue:`805`)
 
 Bug fixes and small improvements:
 
@@ -40,6 +42,7 @@ Bug fixes and small improvements:
 Documentation:
 
 - Fix wrong command in ``How to create a standalone application`` docs (:issue:`792`)
+- Update output html to add github style themes (:issue:`818`)
 
 Internal changes: